Change to match new table behavior [PAR-19]
